Added switch between reprojection/uvtransfer
If using the "Fix Cap Artifacts" feature, the tool remeshes a second time after the displacement happened to get rid of intersecting faces.

In some cases, reprojection the UVs gives better results (if the displacement wasn't too strong), and in others, UV-transfer gives better results (if the remesh wasn't too high-res and using a lot of displacement).

Neither of the methods is absolutely perfect unfortunately